requirements.txt
Django==6.0.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
black==25.12.0
flake8==7.3.0
isort==7.0.0